% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_genome_position.R
\name{adjust_genome_position}
\alias{adjust_genome_position}
\title{Adjust the genome position of target and off-target DArT markers}
\usage{
adjust_genome_position(
  vcf.in,
  alignment,
  reference.name,
  locus.col = "locus_id",
  target.snp.pos.col = "snp_position_in_locus",
  genome.chrom.col = "benlear_v1_chrom",
  locus.start.col = "query_benlear_v1_start",
  locus.end.col = "query_benlear_v1_end",
  genome.start.col = "benlear_v1_start",
  genome.end.col = "benlear_v1_end"
)
}
\arguments{
\item{vcf.in}{A \code{vcfR} object with SNP markers (i.e. from the output of \code{\link{BIGr}[madc2vcf_all]}).}

\item{alignment}{A data frame with information for genome position conversion.}

\item{reference.name}{The name of the reference genome to which positions will be converted.}

\item{locus.col}{The name of the column in \code{alignment} that contains the target locus ID.}

\item{target.snp.pos.col}{The name of the column in \code{alignment} that contains the position of the target SNP in the DArT tag.}

\item{genome.chrom.col}{The name of the column in \code{alignment} that contains the new genome chromosome names.}

\item{locus.start.col}{The name of the column in \code{alignment} that contains the start position of the query tag.}

\item{locus.end.col}{The name of the column in \code{alignment} that contains the end position of the query tag.}

\item{genome.start.col}{The name of the column in \code{alignment} that contains the start position in the genome of the alignment.}

\item{genome.end.col}{The name of the column in \code{alignment} that contains the end position in the genome of the alignment.}
}
\value{
A \code{vcfR} object with SNP positions converted to the position within the genome of interest.
}
\description{
Use BLAST alignment results to modify the genome position of DArT SNP markers.
}
\details{
The function will remove sites that do not have a corresponding position in the new genome.
}
