% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R
\name{as_marker_geno}
\alias{as_marker_geno}
\alias{as_marker_geno.data.frame}
\alias{as_marker_geno.marker.matrix}
\alias{read_geno}
\title{Read or write marker genotype data}
\usage{
as_marker_geno(x, ...)

\method{as_marker_geno}{data.frame}(x, chrom.col, pos.col, cmPos.col = NULL, allele.col = NULL)

\method{as_marker_geno}{marker.matrix}(x, chrom.vec, pos.vec, cmPos.vec = NULL, allele.vec = NULL)

read_geno(
  filename,
  geno,
  ploidy = 2L,
  phased = FALSE,
  sep = ".",
  inbred = FALSE
)
}
\arguments{
\item{x}{A \code{data.frame} or \code{marker.matrix} to be converted into a \code{marker.geno} object.}

\item{...}{Additional arguments to pass to \code{\link[readr]{read_csv}}.}

\item{chrom.col}{The name of the column with chromosome information. (\emph{Required})}

\item{pos.col}{The name of the column with physical chromosome positions. (\emph{Required})}

\item{allele.col}{The name of the column with allele information.}

\item{chrom.vec}{A character vector of chromosome assignments; the length of this vector must equal the number of markers.}

\item{pos.vec}{A numeric vector of positions; the length of this vector must equal the number of markers.}

\item{cmPos.vec}{A numeric vector of genetic positions; the length of this vector must equal the number of markers.}

\item{allele.vec}{A character vector of marker alleles; the length of this vector must equal the number of markers.}

\item{cMpos.col}{The name of the column with centi-Morgan positions.}

\item{file}{Path to the CSV file. The first column must include the marker names; subsequent column names must be marker metadata (chromosom, position, etc.) or sample names.}
}
\description{
Reads marker genotype data to a \code{marker.geno} object, or writes marker genotype data to a file.
}
\section{Functions}{
\itemize{
\item \code{read_geno()}: 

}}
