% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R
\name{as_marker_matrix}
\alias{as_marker_matrix}
\alias{as_marker_matrix.marker.geno}
\alias{as_marker_matrix.matrix}
\title{Convert objects to marker.matrix objects}
\usage{
as_marker_matrix(x, ...)

\method{as_marker_matrix}{marker.geno}(x)

\method{as_marker_matrix}{matrix}(x, chrom.vec, pos.vec, cmPos.vec = NULL, allele.vec = NULL)
}
\arguments{
\item{x}{A \code{marker.geno} object or a matrix of marker genotypes.}

\item{chrom.vec}{A character vector of chromosome assignments; the length of this vector must equal the number of markers.}

\item{pos.vec}{A numeric vector of positions; the length of this vector must equal the number of markers.}

\item{cmPos.vec}{A numeric vector of genetic positions; the length of this vector must equal the number of markers.}

\item{allele.vec}{A character vector of marker alleles; the length of this vector must equal the number of markers.}
}
\value{
A \code{marker.matrix} object.
}
\description{
Convert objects to marker.matrix objects
}
