% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_LD.R
\name{calc_LD_r2}
\alias{calc_LD_r2}
\title{Compute (and plot) linkage disequilibrium decay}
\usage{
calc_LD_r2(
  geno,
  r2.df = NULL,
  by.chrom = TRUE,
  max.pair = 1e+05,
  xlim = c(0, 1e+05),
  method = c("full", "approximate")
)
}
\arguments{
\item{geno}{A data frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{r2.df}{A data frame of pairwise LD estimates, where the first four columns are "chrom", "marker1", "marker2",
and "r2". This can be obtained from the output of \code{calc_LD_r2}.}

\item{max.pair}{The maximum number of marker pairs to use when fitting the LD decay model.}

\item{xlim}{The limits to the x-axis when plotting LD decay.}

\item{method}{The method of computing pairwise LD (r2). If \code{"full"}, complete pairwise correlation coefficients
are computed; if \code{"approximate"}, an VERY approximate matrix multiplication method is used.}
}
\description{
Compute (and plot) linkage disequilibrium decay
}
