% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utils.R
\name{calc_missing}
\alias{calc_missing}
\title{Calculate per-snp or per-genotype missingness}
\usage{
calc_missing(x, type = c("individual", "marker"), check.matrix = TRUE)
}
\arguments{
\item{x}{An n x p marker matrix of n individuals and p markers coded as -1, 0, and 1
for homozygous alternate, heterozygous, and homozygous reference.}

\item{type}{Calculate individual or marker missingness}

\item{check.matrix}{Logical. Should the marker matrix 'x' be checked? Use check.matrix = FALSE
if 'x' contains imputed decimal genotypes.}
}
\description{
Calculate per-snp or per-genotype missingness
}
