% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_duplicates.R
\name{compare_geno}
\alias{compare_geno}
\alias{group_duplicates}
\title{Find and remove highly similar samples based on markers}
\usage{
compare_geno(geno, thresh = 0.95)

group_duplicates(dups)
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals. This can also be the output of \code{compare_geno} in
case you want to set a different threshold without running the marker comparison.}

\item{thresh}{The threshold minimum proportion of matching marker genotypes to return pairs of individuals}

\item{dups}{The output of \code{compare_geno}.}
}
\value{
\itemize{
   \item{compare_geno}: A data frame with pairs of individuals and their similarity (only above the provided threshold)
   \item{group_duplicates}: A numeric vector of group numbers, with element names equal to individuals
 }
}
\description{
Compute the number of matching marker genotypes between pairs of samples. Return the
pairs of individuals with similarity above a threshold.
}
\examples{
geno <- replicate(3000, rbinom(200, 2, 0.5))
dimnames(geno) <- list(paste0("gen", seq_len(nrow(geno))), paste0("marker", seq_len(ncol(geno))))

out <- compare_geno(geno, thresh = 0.4)

# Show potential duplicates
out

# Refilter using a different threshold
out1 <- compare_geno(out, thresh = 0.41)
out1

}
