% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_duplicates.R
\name{drop_duplicates}
\alias{drop_duplicates}
\title{Compare and drop potential duplicates in a marker genotype matrix}
\usage{
drop_duplicates(geno, alias, keep = c("assigned", "lowest.missing"))
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{alias}{A data frame where the first column is the original individual name
and the second column is the new consensus individual name.}

\item{keep}{The preference for keeping a consensus individual. If \code{"assigned"}, the
entry in the first column of \code{alias} will be retained. If \code{"lowest.missing"}, the
entry among those in the second column of \code{alias} with the lowest missing data
proportion will be retained.}
}
\value{
A marker genotype matrix where rows are individuals and columns are markers.
}
\description{
Compare and drop potential duplicates in a marker genotype matrix
}
\details{
For \code{drop_duplicates}, the function will coalesce marker genotypes for samples
with the same consensus name. Individuals not found in \code{alias} will be returned
as is.
}
