% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappoly_functions.R
\name{export_mappoly2_to_diaqtl}
\alias{export_mappoly2_to_diaqtl}
\alias{export_mappoly2_to_qtl2}
\title{Convert haplotype probabilities from mappoly2}
\usage{
export_mappoly2_to_diaqtl(
  data,
  ploidy,
  type = c("genome", "mds"),
  out.prefix = ""
)

export_mappoly2_to_qtl2(data, ploidy, type = c("genome", "mds"))
}
\arguments{
\item{data}{An object of class \code{mappoly2.sequence} (for a single population) or \code{mappoly2.consensus.map}
(for multiple populations). The objects should be the output of \code{calc_haplotypes} or \code{calc_consensus_haplo}.}

\item{ploidy}{The ploidy level}

\item{out.prefix}{The prefix of the output files, including the path to the directory where the files should be saved}
}
\value{
\code{export_mappoly2_to_diaqtl} - Creates a \code{genofile} and \code{pedfile} for import into the \code{diaQTL} package using \code{\link[diaQTL]{read_data}}.

\code{export_mappoly2_to_qtl2} - A list with four elements:
   \itemize{
     \item{\code{probs} - An object of class \code{calc_genoprob} with allele probabilities for
     use in the \code{\link[qtl2]{qtl2}} package.}
     \item{\code{gmap} - A genetic map in list form for use in \code{\link[qtl2]{qtl2}}.}
     \item{\code{pmap} - A physical map in list form for use in \code{\link[qtl2]{qtl2}}.}
     \item{\code{fam.covar} - An incidence matrix for coding family effects when
     conducting a QTL scan in \code{\link[qtl2]{qtl2}}.}
 }
}
\description{
Functions to convert haplotype probabilities from mappoly2 for input into other
packages. Support is included for \code{diaQTL} and \code{qtl2}. The code to convert
to \code{diaQTL} format is borrowed from the \code{\link[diaQTL]{convert_mappoly}} function.
}
\examples{

\dontrun{

alfalfa.f1 <- alfa_f1
alfalfa.f1 <- filter_data(alfalfa.f1, mrk.thresh = 0.2, ind.thresh = 0.1, plot.screening = FALSE)
alfalfa.f1 <- filter_individuals(alfalfa.f1, ind.to.remove = "F1.85.133", inter = FALSE, verbose = FALSE)
alfalfa.f1.all <- pairwise_rf(alfalfa.f1, mrk.scope = "all", ncpus = 8)
alfalfa.f1.all <- rf_filter(alfalfa.f1.all, thresh.LOD.ph = 5, thresh.LOD.rf = 5, thresh.rf = 0.15, probs = c(0.025, 0.975), diagnostic.plot = FALSE)
g <- group(x = alfalfa.f1.all, expected.groups = 8, comp.mat = TRUE, inter = FALSE)
s <- make_sequence(g, ch = list(1, 2, 3, 4, 5, 6, 7, 8))
s <- order_sequence(s, type = "genome")
s <- rf_filter(s, type = "genome", probs = c(0.025, 0.975), diag.markers = 50)
s <- pairwise_phasing(s, type = "genome", thresh.LOD.ph = 3, thresh.LOD.rf = 3, thresh.rf = 0.5, max.search.expansion.p1 = 10, max.search.expansion.p2 = 10)
s <- mapping(s, type = "genome", parent = "p1", ncpus = 8)
s <- mapping(s, type = "genome", parent = "p2", ncpus = 8)
s <- merge_single_parent_maps(s, type = "genome", ncpus = 8, error = 0.05)

s <- calc_haplotypes(s, type = "genome", ncpus = 8)
s <- augment_phased_map(s, type = "genome", ncpus = 8)
s <- calc_haplotypes(s, type = "genome", ncpus = 8)
I195xJ432_map <- s

# Load the second map
download.file("https://github.com/mmollina/mappoly2_vignettes/raw/main/I195_x_F1-85-209_map.rda",
destfile = "temp_file.rda")

load("temp_file.rda")

I195xF1_85_209_map <- mapping(I195xF1_85_209_map, type = "genome", error = 0.05, ncpus = 8)

# Integrate the two maps
MAPs <- list(I195xJ432 = I195xJ432_map, I195xF1_85_209 = I195xF1_85_209_map)
prep.maps <- prepare_to_integrate(MAPs)

consensus_map <- estimate_consensus_map(prep.maps, ncpus = 8, err = 0.05)
consensus_map <- calc_consensus_haplo(consensus_map, ncpus = 8)



# Export to diaQTL
export_mappoly2_to_diaqtl(data = I195xJ432_map, ploidy = 4, out.prefix = "alfalfa_gp_")
export_mappoly2_to_diaqtl(data = consensus_map, ploidy = 4, out.prefix = "alfalfa_gp_consensus_")

# Read into diaQTL
dat1 <- diaQTL::read_data(genofile = "alfalfa_gp_diaQTL_geno.csv", ploidy = 4, pedfile = "alfalfa_gp_diaQTL_ped.csv")
dat2 <- diaQTL::read_data(genofile = "alfalfa_gp_consensus_diaQTL_geno.csv", ploidy = 4, pedfile = "alfalfa_gp_consensus_diaQTL_ped.csv")

# Export to qtl2
qtl2_dat1 <- export_mappoly2_to_qtl2(data = I195xJ432_map, ploidy = 4, type = "genome")
qtl2_dat2 <- export_mappoly2_to_qtl2(data = consensus_map, ploidy = 4, type = "genome")

# Use in qtl2
kin1 <- qtl2::calc_kinship(probs = qtl2_dat2$probs, type = "loco")
pheno <- c(MAPs$I195xJ432$data$geno.dose['Alf_2262657_SNP1',] * 2,
           MAPs$I195xF1_85_209$data$geno.dose['Alf_2262657_SNP1',] * 2)
pheno <- as.matrix(pheno)

ans <- qtl2::scan1(genoprobs = qtl2_dat2$probs, pheno = pheno, kinship = kin1, addcovar = qtl2_dat2$fam.covar)
plot(ans, map = qtl2_dat2$pmap)

}

}
