% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geno.R, R/marker_genotypes.R
\name{filter_geno}
\alias{filter_geno}
\title{Filter SNP marker genotypes}
\usage{
filter_geno(
  x,
  marker.summary,
  max.marker.miss = 1,
  max.sample.miss = 1,
  range.marker.het = c(0, 1),
  range.sample.het = c(0, 1),
  min.maf = 0,
  markers,
  samples,
  verbose = TRUE
)

filter_geno(
  x,
  marker.summary,
  max.marker.miss = 1,
  max.sample.miss = 1,
  range.marker.het = c(0, 1),
  range.sample.het = c(0, 1),
  min.maf = 0,
  markers,
  samples,
  verbose = TRUE
)
}
\arguments{
\item{x}{A \code{marker.matrix} or \code{marker.geno} object.}

\item{marker.summary}{An optional marker summary object for x. If passed, it may speed up the filtering function.}

\item{max.marker.miss}{The maximum missingness rate for a marker.}

\item{max.sample.miss}{The maximum missingness rate for a sample.}

\item{range.marker.het}{A length-two vector of the minimum and maximum marker heterozygosity rate.}

\item{range.sample.het}{A length-two vector of the minimum and maximum sample heterozygosity rate.}

\item{min.maf}{The minimum minor allele frequency for a marker.}

\item{markers}{A character vector of markers to select. If passed, can be \code{"all"} or a character vector of markers to subset; in this case, all other filtering criteria are ignored.}

\item{samples}{A character vector of samples to select. If passed, can be \code{"all"} or a character vector of samples to subset; in this case, all other filtering criteria are ignored.}

\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{max.ind.miss}{The maximum rate of missingness to retain an individual}

\item{max.mar.miss}{The maximum rate of missingness to retain an marker}

\item{min.mac}{The minimum minor allele count to retain a marker.}

\item{max.r2}{The maximum squared correlation between any pair of markers. The \code{r^2} is used directly in filtering. For each pair of markers exceeding this threshold, the one with the greater minor allele frequency is retained. Set \code{max.r2 = 1} for no filtering.}
}
\value{
Variable of class \code{\link{geno}}.
}
\description{
Filter SNP marker genotypes

Filter marker genotype data
}
\examples{
# Use data from the PopVar package
data("cranberry_geno", package = "PopVar")

# Remove cM, rename
geno <- cranberry_geno[-3]
names(geno)[1:3] <- c("marker", "chrom", "position")

# Filter
geno_in <- filter_geno

}
