% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_madc.R
\name{filter_madc}
\alias{filter_madc}
\title{Filter MADC data}
\usage{
filter_madc(
  madc,
  madc.file,
  min.mh.reads = 2,
  min.mh.samples = 10,
  min.mh.ref.identity = 0.9,
  min.mh.ref.coverage = 0.9,
  min.locus.samples = 10,
  min.locus.reads = 5,
  max.locus.haps = 10,
  max.sample.missing = 0.95
)
}
\arguments{
\item{madc}{A data.frame of MADC data (read from a .csv); ignored if \code{madc.file} is passed.}

\item{madc.file}{The name of the MADC file; ignored if \code{madc} is passed.}

\item{min.mh.reads}{Minimum number of reads at genotyped samples to retain a RefMatch or AltMatch allele.}

\item{min.mh.samples}{Minimum number of samples required to retain a RefMatch or AltMatch allele.
has at least \code{min.mh.reads} reads.}

\item{min.mh.ref.identity}{Minimum proportion of identity of a RefMatch or AltMatch allele to the reference.}

\item{min.mh.ref.coverage}{Minimum proportion of coverage of a RefMatch or AltMatch allele to the reference.}

\item{min.locus.samples}{Minimum number of samples to retain a locus.}

\item{min.locus.reads}{Minimum number of reads at a genotyped locus to retain a locus.}

\item{max.locus.haps}{Maximum number of haplotypes at a locus. If greater, only the Ref and Alt haplotypes are retained.}

\item{max.sample.missing}{Maximum missing data proportion to retain a sample/individual.}
}
\description{
Applies filters to MADC data from DArTag experiments
}
