% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_vcf.R
\name{filter_vcf}
\alias{filter_vcf}
\title{Filter genotype calls in a VCF object}
\usage{
filter_vcf(x, minAD = NULL, minQ = NULL, remove.completely.missing = FALSE)
}
\arguments{
\item{x}{A \code{vcfR} object.}

\item{minAD}{The minimum allele depth to retain a genotype call; those below this depth are set to missing.}

\item{minQ}{The minimum site quality score to retain a site.}

\item{remove.completely.missing}{Logical. Should completely missing genotypes be removed?}
}
\description{
Applies filters to genotype calls and sites in a VCF file. The function is designed to apply
allele depth filters to heterozygous genotype calls.
}
