% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_duplicates.R
\name{find_alias_geno_missing}
\alias{find_alias_geno_missing}
\title{Reduce a marker matrix by duplicates}
\usage{
find_alias_geno_missing(geno, alias, sep = ", ")
}
\arguments{
\item{geno}{A marker genotype matrix, where rows are markers and columns are individuals.}

\item{alias}{A data frame of individual and alias information, where the first column
is the desired genotype ID, and the second column contains alias information, if applicable}

\item{sep}{The delimiter separating multiple aliases in the second column of \code{alias}.}
}
\description{
Takes a marker matrix and a data frame of geno_ids and aliases
and returns the geno_id or alias with the lowest missing marker data rate
}
