% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_find_genes.R
\name{find_genes}
\alias{find_genes}
\title{Find genes from a table of markers}
\usage{
find_genes(qtl, genes, window = 20000)
}
\arguments{
\item{qtl}{A data frame with the name, chromosome, and physical position of markers; one row per marker.}

\item{genes}{A \code{GRanges} object with gene model data for the same reference genome as the position of the markers in \code{qtl}.}

\item{window}{Numeric size of the window from each marker in \code{qtl} to find genes.}
}
\description{
Find genes based on a table of markers, such as significant QTL peaks from a mapping
study or GWAS, or outlier SNPs from other analyses.
}
