% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_snp.R
\name{find_snp}
\alias{find_snp}
\title{Find the SNP position between two sequences}
\usage{
find_snp(seq1, seq2, unambiguous = FALSE)
}
\arguments{
\item{seq1}{A sequence}

\item{seq2}{Another sequence}

\item{unambiguous}{Logical. Use only unambiguous nucleotides.}
}
\description{
This is a naive function to find nucleotide mismatches between two sequences.
}
