% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geno.LD-class}
\alias{geno.LD-class}
\alias{geno.LD}
\title{S4 class of marker genotype data with linkage disequilibrium information}
\description{
S4 class of marker genotype data with linkage disequilibrium information
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}

\item{\code{LD}}{A matrix of pairwise marker LD estimates}
}}

