% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geno.mh-class}
\alias{geno.mh-class}
\alias{geno}
\title{S4 class of SNP genotype data}
\description{
S4 class of SNP genotype data

S4 class of microhaplotype genotype data
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}

\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}
}}

