% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geno.relmat-class}
\alias{geno.relmat-class}
\alias{geno.relmat}
\title{S4 class of marker genotype data with genomic relationship matrix information}
\description{
S4 class of marker genotype data with genomic relationship matrix information
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{map}}{A data frame of the genetic map information.}

\item{\code{geno.mat}}{A matrix of marker genotypes.}

\item{\code{haplo.mat}}{A matrix of haplotypes.}

\item{\code{phased}}{Whether the marker data was phased.}

\item{\code{idx}}{The indices of individuals retained after filtering.}

\item{\code{idy}}{The indices of markers retained after filtering.}

\item{\code{G}}{The additive genomic relationship matrix.}

\item{\code{D}}{The dominance relationship matrix.}
}}

