% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{geno.summary-class}
\alias{geno.summary-class}
\alias{geno.summary}
\title{S4 class of a summary of a marker genotype object}
\description{
S4 class of a summary of a marker genotype object
}
\section{Slots}{

\describe{
\item{\code{ploidy}}{The ploidy level.}

\item{\code{nMar}}{The number of markers.}

\item{\code{nInd}}{The number of individuals.}

\item{\code{mac}}{The minor allele count at each locus.}

\item{\code{mar.missing}}{The proportion of missing data at each marker locus.}

\item{\code{ind.missing}}{The proportion of missign data of each individual}
}}

