% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geno_PCA.R
\name{geno_PCA}
\alias{geno_PCA}
\title{Principal component analysis of marker genotypes}
\usage{
geno_PCA(
  geno,
  metadata,
  input = c("relmat", "geno.mat"),
  ploidy = 2,
  n.core = 1
)
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals. The \code{geno} object is assumed to be filtered.}

\item{metadata}{Metadata for the individuals in \code{geno}. Used for plotting. The first column
of \code{metadata} must be "geno_id" for the individual name. Other columns are metadata columns.}

\item{input}{What should be the input for PCA? Can be \code{relmat} and an additive relationship
matrix will be computed, or \code{geno.mat} and PCA will be performed on the marker genotype matrix.}

\item{ploidy}{Integer ploidy number.}

\item{n.core}{The number of cores for processing.}
}
\description{
Principal component analysis of marker genotypes
}
