% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_support_functions.R
\name{geno_prob}
\alias{geno_prob}
\title{Genotype probabilities given parent markers and the generation of selfing}
\usage{
geno_prob(f.gen, r)
}
\arguments{
\item{f.gen}{The filial (F) generation (e.g. if \code{f.gen} is 3, this refers to an F_3 family).}

\item{r}{The recombination frequency. If passed, two-locus genotype probabilities are provided.}
}
\value{
A \code{matrix} of conditional genotype probabilities for offspring marker genotypes (-1, 0, 1) (columns) given the marker genotypes of two parents (rows).
}
\description{
Genotype probabilities given parent markers and the generation of selfing
}
