% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenomicBreeding-package.R
\docType{package}
\name{GenomicBreeding-package}
\alias{GenomicBreeding}
\alias{GenomicBreeding-package}
\title{GenomicBreeding: Tools for genomics-assisted breeding in plants}
\description{
Functionality for genomics-assisted breeding in plants, including marker manipulation, genomewide prediction, and cross-validation.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
