% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_genotypes.R
\name{impute_geno}
\alias{impute_geno}
\title{Impute markers using rrBLUP}
\usage{
impute_geno(
  geno,
  ploidy,
  method = c("pop", "EM", "RF"),
  ind.max.missing,
  snp.max.missing,
  params = NULL,
  n.core = 1
)
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{ploidy}{Integer ploidy number.}

\item{method}{Imputation method: the population mean ("pop"), expectation-maximization ("EM"), or random forest ("RF").}

\item{ind.max.missing}{Maximum missingness threshold for individuals.}

\item{snp.max.missing}{Maximum missingness threshold for markers.}

\item{params}{A \code{list} of parameters.}

\item{n.core}{The number of cores for processing.}
}
\description{
Impute markers using rrBLUP. This function is borrowed from the \code{polyBreedR} package.
}
