% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_utils.R
\name{impute_linkage}
\alias{impute_linkage}
\title{Internal functions}
\usage{
impute_linkage(hd.file, ld.file, out.file, hd.geno, ld.geno, ped, n.core = 1)
}
\description{
Internal package documents or those without clear documentation.
}
