% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_genotypes.R
\name{impute_ref}
\alias{impute_ref}
\title{Create a reference set of haplotypes from a high-density set}
\usage{
impute_ref(
  in.file,
  out.file,
  mar.max.missing,
  ind.max.missing,
  method = c("beagle", "api2"),
  keep.ind,
  keep.mar,
  params,
  beagle.path,
  n.core = 1
)
}
\arguments{
\item{in.file}{Input VCF file.}

\item{out.file}{Output VCF file.}

\item{mar.max.missing}{Maximum missingness threshold for markers.}

\item{ind.max.missing}{Maximum missingness threshold for individuals.}

\item{method}{Imputation and phasing method ("beagle" for Beagle, "api2" for AlphaPlantImpute2)}

\item{keep.ind}{A vector of individuals to keep in the VCF file.}

\item{keep.mar}{A vector of markers to keep in the VCF file.}

\item{params}{A \code{list} of parameters for beagle.}

\item{beagle.path}{Path to the executable file for beagle.}

\item{n.core}{The number of cores for processing.}
}
\description{
Create a reference set of haplotypes from a high-density set
}
