% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_genotypes.R
\name{impute_ref2low}
\alias{impute_ref2low}
\title{Impute and phase a low-density offspring set using a high-density phased reference set}
\usage{
impute_ref2low(
  in.file,
  out.file,
  ref.file,
  ped.file,
  ind.max.missing,
  beagle.params,
  api2.params,
  update.ref = TRUE,
  n.core = 1
)
}
\arguments{
\item{in.file}{Input VCF file.}

\item{out.file}{Output VCF file.}

\item{ref.file}{The phased reference VCF file.}

\item{ped.file}{The pedigree file or a ped object.}

\item{ind.max.missing}{Maximum missingness threshold for individuals.}

\item{beagle.params}{A \code{list} of parameters for beagle.}

\item{api2.params}{A \code{list} of parameters for AlphaPlantImpute2.}

\item{update.ref}{Logical. Should the reference set be updated with the low-density phased haplotypes?}

\item{n.core}{The number of cores for processing.}
}
\description{
Impute and phase a low-density offspring set using a high-density phased reference set
}
