% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_haplotypes_polyorigin.R
\name{infer_haplotypes_polyorigin}
\alias{infer_haplotypes_polyorigin}
\title{Reconstruct haplotypes using PolyOrigin}
\usage{
infer_haplotypes_polyorigin(
  geno,
  ped,
  out.prefix,
  ploidy = 2,
  filter.mars = FALSE,
  max.ind.miss = 0.2,
  max.mar.miss = 0.2,
  julia.path = ""
)
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "cM" (genetic position);
optionally, the fourth column can be the physical position with column name "bp". Subsequent columns are genotype calls for individuals.}

\item{ped}{A pedigree data frame where the first four columns are the individual name, the family (or population) name,
the mother genotype ID, and the father genotype ID.}

\item{out.prefix}{The prefix (i.e. no file extension) for saving the output of PolyOrigin.}

\item{ploidy}{Ploidy level; must be 2, 4, or 6.}

\item{filter.mars}{Logical; should the data for each family be filtered using mappoly?}

\item{max.ind.miss}{The maximum rate of missingness to retain an individual}

\item{max.mar.miss}{The maximum rate of missingness to retain an marker}

\item{julia.path}{The path to the Julia executable on the system.}
}
\description{
Reconstruct haplotypes using PolyOrigin
}
