% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_support_functions.R
\name{interpolate_positions}
\alias{interpolate_positions}
\title{Interpolate marker genetic positions}
\usage{
interpolate_positions(x, reference, max.recomb.rate = 7)
}
\arguments{
\item{reference}{A data.frame of physical map intervals with smoothed recombination rates. Columns should be: 1) chromosome, 2) left physical position boundary, 3) right physical position boundary, and 4) recombination rate (in cM/Mb).}

\item{max.recomb.rate}{The maximum value of smoothed recombination rates. Any values in \code{reference} that exceed this value will be converted to this value.}

\item{data}{A data.frame of SNPs with positions to be extrapolated. This should resemble the table format of genetic maps from the R/qtl package: first column is chromosome, second column is physical position, row.names are marker names.}
}
\description{
Uses a reference map of recombination rates (cM/Mb) to interpolate the genetic position of markers based on their physical positions.
}
