% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_find_genes.R
\name{load_cran_gff}
\alias{load_cran_gff}
\title{Load the gene models for the cranberry Ben Lear reference genome}
\usage{
load_cran_gff(path = NULL, download = FALSE)
}
\arguments{
\item{path}{A path to the GFF file (or url). If NULL (default), the raw GFF will be downloaded from \link{vaccinium.org}.}

\item{download}{Should the GFF file be downloaded?}
}
\description{
Load the gene models for the cranberry Ben Lear reference genome
}
