% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_calling.R
\name{madc2updog}
\alias{madc2updog}
\title{Convert a MADC file to objects for genotyping calling in updog}
\usage{
madc2updog(madc, botloci, target.only = FALSE)
}
\arguments{
\item{madc}{A data.frame of MADC data (read from a .csv).}

\item{botloci}{A character vector indicating the names of loci for which the sequence is on the bottom strand with respect to the reference genome.
See \code{\link[polyRAD]{readDArTag}}.}

\item{target.only}{Logical. Should only target reference and alternate tags be retained? If TRUE, the output data is bi-allelic SNPs only; if FALSE,
the output data is multi-allelic loci.}
}
\description{
Convert a MADC file to objects for genotyping calling in updog
}
\details{
This function uses polyRAD to import the microhaplotype data, but converts it
to updog format for genotyping calling
}
