% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vcfs.R
\name{merge_count_vcfs}
\alias{merge_count_vcfs}
\title{Merge VCF files using allele depth information only}
\usage{
merge_count_vcfs(file.list, run.ids = NULL, n.core = 1)
}
\arguments{
\item{file.list}{A character vector of VCF file paths}

\item{run.ids}{Vector of character defining the run ID for each file}

\item{n.core}{Number of processors for parallel computation.}
}
\description{
Merge VCF files using allele depth information only
}
