% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotype_duplicates.R
\name{merge_duplicates}
\alias{merge_duplicates}
\title{Merge duplicates in a marker genotype matrix}
\usage{
merge_duplicates(geno, alias, snps.use, run.ids = NULL, min.concordance = 0.9)
}
\arguments{
\item{geno}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{alias}{A data frame where the first column is the original individual name
and the second column is the new consensus individual name.}

\item{snps.use}{The SNPs to be used to compare duplicates. Defaults to all markers.}

\item{min.concordance}{The minimum percentage of concordance among non-missing genotype calls among
duplicate samples in order to merge those samples. If the observed concordance is less than
\code{min.concordance}, those samples will not be merged and a warning will be
printed. Set this to a low number to ensure merging.}
}
\value{
A marker genotype matrix where rows are individuals and columns are markers.
}
\description{
Merge duplicates in a marker genotype matrix
}
\details{
Individuals not found in \code{alias} will be returned as is.
}
