% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_parent_offspring_testing.R
\name{parent_offspring_test_mf}
\alias{parent_offspring_test_mf}
\alias{parent_offspring_test}
\title{Conduct marker quality control using parent-offspring tests}
\usage{
parent_offspring_test_mf(
  x,
  family.list,
  min.fam.size = 5,
  correct = TRUE,
  mismatch.tol = 3
)

parent_offspring_test(
  parents,
  offspring,
  f.gen,
  correct = TRUE,
  min.fam.size = 5,
  mismatch.tol = 3
)
}
\arguments{
\item{x}{A \code{marker.matrix} object.}

\item{family.list}{A list of data frames that each include the columns "sample" (sample name), "pedigree", "parent1", "parent2", and "f_gen" (inbreeding generation). If the elements also contain the columns "parent1_name" and "parent2_name", missing parent names will be converted to these names.}

\item{min.fam.size}{The minimum family size necessary to correct offspring/parent genotype states when offspring are polymorphic.}

\item{correct}{Logical: should marker genotypes be corrected? Correction only happens when possible.}

\item{parents}{A 2 x m matrix of parent marker genotypes.}

\item{offspring}{A n x m matrix of offspring marker genotypes.}

\item{f.gen}{The filial (inbreeding) generation of offspring in the family (i.e. \code{f.gen = 3} = F_3 offspring).}
}
\description{
Uses heuristics and expected genotype probabilities to determine aberrant markers and (optionally) correct them.
}
\section{Functions}{
\itemize{
\item \code{parent_offspring_test()}: Conduct a parent-offpring test on a single family.

}}
