% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qtl_model.R
\name{plot_fit1}
\alias{plot_fit1}
\title{Plot the results of fitted QTL model}
\usage{
plot_fit1(fit1out, genoprobs)
}
\arguments{
\item{fit1out}{The output from \code{\link{fit1}[qtl]}.}

\item{genoprobs}{Genotype probabilities.}
}
\description{
Plot the results of fitted QTL model
}
