% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_haplotypes_polyorigin.R
\name{polyorigin_to_rqtl2}
\alias{polyorigin_to_rqtl2}
\title{Convert the output of polyorigin to a genoprobs object for R/qtl2}
\usage{
polyorigin_to_rqtl2(genofile, pedfile, ploidy = 2)
}
\arguments{
\item{genofile}{Path to the _genofile.csv output from polyorigin haplotype reconstruction}

\item{pedfile}{Path to the _pedfile.csv output from polyorigin haplotype reconstruction}

\item{ploidy}{The ploidy level}
}
\description{
Convert the output of polyorigin to a genoprobs object for R/qtl2
}
