% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qtl_model.R
\name{prep_cross2}
\alias{prep_cross2}
\title{Prep a qtl2 cross object for QTL modelling}
\usage{
prep_cross2(
  cross,
  ped,
  ploidy = 2,
  error.prob = 0.05,
  cores = 1,
  crosstype = c("f2", "outcross"),
  probs = NULL,
  map = NULL
)
}
\description{
Prep a qtl2 cross object for QTL modelling
}
