% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print,geno.summary-method}
\alias{print,geno.summary-method}
\title{Summarize a \code{geno} object (and derived objects)}
\usage{
\S4method{print}{geno.summary}(x)
}
\description{
Summarize a \code{geno} object (and derived objects)
}
