% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R
\name{prune_LD_geno}
\alias{prune_LD_geno}
\title{Prune SNPs based on linkage disequilibrium (LD)}
\usage{
prune_LD_geno(x, maxLD = 0.95, posWindowSize, snpWindowSize, verbose = FALSE)
}
\arguments{
\item{x}{A \code{marker.geno} object.}

\item{maxLD}{The maximum LD of SNPs within a window.}

\item{posWindowSize}{The size of the sliding window (in bp).}

\item{snpWindowSize}{The maximum number of SNPs in the sliding window.}

\item{verbose}{Logical; should the function print status messages?}
}
\value{
A \code{marker.geno} object with SNPs filtered according to LD.
}
\description{
Prune SNPs based on linkage disequilibrium (LD)
}
