% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappoly_functions.R
\name{pull_consensus_map}
\alias{pull_consensus_map}
\title{Pull the consensus map}
\usage{
pull_consensus_map(
  x,
  lg,
  type = c("genetic", "physical", "both"),
  format = c("table", "qtl2")
)
}
\arguments{
\item{x}{A \code{mappoly2.consensus.map} object.}

\item{lg}{The linkage groups to pull (defaults to all).}

\item{type}{The type of map to return (can be genetic, physical, or both).}

\item{format}{The format of the output (can be a table or list format for qtl2).}
}
\description{
Pull the consensus map
}
