% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_geno.R
\name{read_geno}
\alias{read_geno}
\title{Read in SNP marker genotype data}
\usage{
read_geno(
  filename,
  geno,
  ploidy = 2L,
  phased = FALSE,
  sep = ".",
  inbred = FALSE
)
}
\arguments{
\item{filename}{Name of CSV file with marker allele dosage}

\item{geno}{A data frame of marker allele dosages. If passed, \code{filename} is ignored.}

\item{ploidy}{The ploidy level e.g. 2, 4, 6, ..., etc.}

\item{phased}{If TRUE, the input genotype matrix is phased, with alleles coded as positive integers. The name of each
individual/clone column is the id and haplotype, concatenated by \code{sep}.}

\item{sep}{The haplotype concatenator; see \code{phased}.}

\item{all.hom}{Logical. If unphased marker data is passed, should the function assume that all individuals are completely homozygous?}
}
\value{
Variable of class \code{\link{geno}}.
}
\description{
Read in SNP marker genotype data
}
\details{
The first columns of the file or geno object should be marker, chrom, cM (and optionally bp, if provided). Subsequent columns
contain the allele dosage for individuals/clones, coded 0,1,2. While fractional values are allowed for genomewide prediction,
these values are rounded when predicting the genetic variance within crosses.
}
\examples{
# Use data from the PopVar package
data("barley_geno", package = "PopVar")
data("cranberry_geno", package = "PopVar")
data("cranberry_geno_phased", package = "PopVar")

# Read in barley marker genotypes assuming no phasing and non-inbred lines.
geno_in <- read_geno(geno = barley_geno)

# Read in barley marker genotypes assuming completely inbred lines
geno_in <- read_geno(geno = barley_geno, inbred = TRUE)

# Read in cranberry marker genotype assuming no phasing
geno_in <- read_geno(geno = cranberry_geno)

# Read in cranberry phased haplotypes from Beagle
geno_in <- read_geno(geno = cranberry_geno_phased, phased = TRUE, sep = "_hap")

# Simulated polyploidy data
geno <- replicate(1000, rbinom(n = 100, size = 4, prob = runif(1)))
geno <- t(geno)
colnames(geno) <- paste0("id", 1:100)
meta <- data.frame(marker = paste0("m", 1:1000), chrom = 1, cM = seq(0, 100, length.out = 1000))
geno <- cbind(meta, geno)
geno_in <- read_geno(geno = geno, ploidy = 4)

}
