% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removed_coincident_sites.R
\name{remove_coincident_sites}
\alias{remove_coincident_sites}
\title{Remove SNPs with duplicate positions in a VCF}
\usage{
remove_coincident_sites(x)
}
\arguments{
\item{x}{An \code{vcfR} object.}
}
\description{
Remove SNPs with duplicate positions in a VCF
}
