% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R
\name{resolve_duplicate_geno}
\alias{resolve_duplicate_geno}
\alias{resolve_duplicate_geno.marker.geno}
\alias{resolve_duplicate_geno.marker.matrix}
\title{Resolve duplicate entries in geno objects}
\usage{
resolve_duplicate_geno(x)

\method{resolve_duplicate_geno}{marker.geno}(x)

\method{resolve_duplicate_geno}{marker.matrix}(x)
}
\arguments{
\item{x}{A \code{marker.geno} or \code{marker.matrix} object with duplicates.}
}
\description{
Resolve duplicate entries in geno objects
}
\details{
Duplicates are resolved by using the consensus genotype (i.e. the most frequent genotype at a marker across duplicate samples). If more than 2 marker states are present among the duplicate samples, the genotype call is set to missing (i.e. NA).
}
