% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_marker_genotypes.R
\name{rrblup2ag}
\alias{rrblup2ag}
\title{Convert a rrBLUP formatted object to AlphaGenes format}
\usage{
rrblup2ag(
  geno.df,
  outdir = ".",
  sep.by.chrom = TRUE,
  prefix = "cran_snps_alphaplantimpute_input"
)
}
\arguments{
\item{geno.df}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{outdir}{The path the output directory to save the files.}

\item{sep.by.chrom}{Logical. Should markers be separate by chromosome?}

\item{prefix}{The filename prefix}
}
\description{
Convert a rrBLUP formatted object to AlphaGenes format
}
