% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_marker_genotypes.R
\name{rrblup2genomat}
\alias{rrblup2genomat}
\title{Convert rrblup formatted genotype data to a geno matrix}
\usage{
rrblup2genomat(x, transpose = TRUE, output = c("012", "-101"))
}
\arguments{
\item{x}{A data.frame of genotype calls, where the first three columns are "marker," "chrom,", and "position."
Subsequent columns are genotype calls for individuals.}

\item{transpose}{Logical. Should the output matrix be indiv x markers (TRUE) or markers x indiv (FALSE)}

\item{output}{The output type, either 0-1-2 for homozygous reference, heterozygous, homozygous alternate, or -1-0-1 for the same.}
}
\description{
Convert rrblup formatted genotype data to a geno matrix
}
