% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_calling.R
\name{run_updog}
\alias{run_updog}
\title{Call marker genotypes using updog}
\usage{
run_updog(x, ploidy = 2, model = "norm", n.cores = 1)
}
\arguments{
\item{x}{A list of two matrices from the function \code{\link{madc2updog}}.}

\item{ploidy}{The ploidy of all individuals in the population.}

\item{model}{See \code{\link[updog]{multidog}}.}

\item{n.cores}{Number of CPU cores for parallel processing.}
}
\description{
This is a simple wrapper function around functions from \code{\link[updog]{updog}} but
is made aware of multi-allelic markers that are not handled natively by updog.
}
