% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_summaries.R
\name{summarize_geno}
\alias{summarize_geno}
\title{Summarize a \code{geno} object (and derived objects)}
\usage{
summarize_geno(geno)
}
\arguments{
\item{geno}{A \code{geno} object (or its derivatives).}
}
\description{
Summarize a \code{geno} object (and derived objects)
}
