% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R
\name{summary_geno}
\alias{summary_geno}
\alias{summary_geno.marker.matrix}
\alias{summary_geno.marker.geno}
\alias{print.marker.summary}
\alias{print.marker.geno}
\alias{print.marker.matrix}
\alias{plot.marker.summary}
\title{Summarize marker genotype data}
\usage{
summary_geno(x, simple = FALSE)

\method{summary_geno}{marker.matrix}(x, simple = FALSE)

\method{summary_geno}{marker.geno}(x, simple = FALSE)

print.marker.summary(x)

print.marker.geno(x)

print.marker.matrix(x)

plot.marker.summary(x)
}
\arguments{
\item{x}{A \code{marker.matrix} object or \code{marker.geno} object.}

\item{simple}{Logical; should the summary be simple (i.e. just the number of markers and samples)? If FALSE, a full summary is calculated.}
}
\description{
Summarize marker genotype data
}
\details{
Calculates the number of markers, number of samples, marker and sample missingness,
marker and sample heterozygosity, marker allele frequency, and minor allele frequency.
}
