% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_marker_genotypes.R
\name{vcf2ag}
\alias{vcf2ag}
\title{Convert a VCF object to AlphaGenes format}
\usage{
vcf2ag(vcf, outdir = ".", sep.by.chrom = TRUE, n.core = 1)
}
\arguments{
\item{vcf}{A \code{vcfR} object.}

\item{outdir}{The path the output directory to save the files.}

\item{sep.by.chrom}{Logical. Should markers be separate by chromosome?}

\item{n.core}{The number of computer cores to use.}
}
\description{
Convert a VCF object to AlphaGenes format
}
