% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotype_calling.R
\name{vcf2updog}
\alias{vcf2updog}
\title{Convert a VCF to read count matrices for updog}
\usage{
vcf2updog(vcf.in, variants.keep = NULL, keep.multiallelic = FALSE)
}
\arguments{
\item{vcf.in}{A \code{vcfR} object.}

\item{variants.keep}{A vector of variant IDs to keep for output. If \code{NULL}, all variants are kept.}

\item{keep.multiallelic}{If \code{TRUE}, multi-allelic SNPs are kept, but only the most frequent alternate allele is retained. If \code{FALSE}, multi-allelic SNPs are removed.}
}
\value{
A list with two matrices: \code{refmat} (the reference allele read count matrix) and \code{sizemat} (the total allele read count matrix).
}
\description{
Convert a VCF to read count matrices for updog
}
