% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_geno.R
\name{filter_geno}
\alias{filter_geno}
\title{Filter SNP marker genotypes}
\usage{
filter_geno(
  geno,
  max.ind.miss = 0.5,
  max.mar.miss = 0.5,
  min.mac = 5,
  max.r2 = 1
)
}
\arguments{
\item{geno}{A object of \class{geno} (i.e. output of function \code{\link{read_geno}}).}

\item{max.ind.miss}{The maximum rate of missingness to retain an individual}

\item{max.mar.miss}{The maximum rate of missingness to retain an marker}

\item{min.mac}{The minimum minor allele count to retain a marker.}

\item{max.r2}{The maximum squared correlation between any pair of markers. The \code{r^2} is used directly in filtering. For each pair of markers exceeding this threshold, the one with the greater minor allele frequency is retained. Set \code{max.r2 = 1} for no filtering.}
}
\value{
Variable of class \code{\link{geno}}.
}
\description{
Filter SNP marker genotypes
}
\examples{
# Use data from the PopVar package
data("barley_geno", package = "PopVar")
data("cranberry_geno", package = "PopVar")
data("cranberry_geno_phased", package = "PopVar")

# Read in marker genotype data
geno_in <- read_geno(geno = barley_geno)

# Filter
geno_in <- filter_geno

}
