% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_imputation.R
\name{impute_geno_flanking_mf}
\alias{impute_geno_flanking_mf}
\alias{impute_geno_flanking}
\alias{impute_geno_em}
\title{Impute marker genotypes in families using flanking markers}
\usage{
impute_geno_flanking_mf(
  x,
  family.list,
  marker.map,
  min.geno.prob = 0.7,
  map.function = c("haldane", "kosambi", "c-f", "morgan"),
  return.geno.prob = FALSE
)

impute_geno_flanking(
  parents,
  offspring,
  marker.map,
  f.gen,
  min.geno.prob = 0.7,
  map.function = c("haldane", "kosambi", "c-f", "morgan"),
  return.geno.prob = FALSE
)

impute_geno_em(x, method = c("EM", "mean"), verbose = TRUE)
}
\arguments{
\item{x}{A \code{marker.matrix} object.}

\item{family.list}{A list of data frames that each include the columns "sample" (sample name), "pedigree", "parent1", "parent2", and "f_gen" (inbreeding generation). If the elements also contain the columns "parent1_name" and "parent2_name", missing parent names will be converted to these names.}

\item{marker.map}{A data frame of marker map positions, where the first column is the marker name, the second column is the chromosome, and the third column is the genetic map position.}

\item{min.geno.prob}{The minimum marker genotype probability needed for imputation.}

\item{map.function}{The mapping function to convert cM distance to recombination frequency.}

\item{return.geno.prob}{Logical; should genotype probabilities also be returned?}

\item{parents}{A 2 x m matrix of parent marker genotypes.}

\item{offspring}{A n x m matrix of offspring marker genotypes.}

\item{f.gen}{The filial (inbreeding) generation of offspring in the family (i.e. f.gen = 3 = F_3 offspring).}

\item{method}{The imputation method; may be "EM" for expectation-maximization or "mean" for the marker mean.}

\item{verbose}{Logical; should status messages be printed?}
}
\value{
A list containing i) the new marker matrix with imputed marker genotypes and ii) the missingness statistics for pre-  and post-imputation.

A matrix of imputed offspring marker genotypes.
}
\description{
Imputed marker genotypes for multiple families using flanking markers and recombination frequencies.
}
\section{Functions}{
\itemize{
\item \code{impute_geno_flanking()}: Impute marker genotypes in a single family using flanking markers

\item \code{impute_geno_em()}: Impute marker genotypes using an expectation-maximization algorithm or the marker mean; implements this one chromosome at-a-time.

}}
