% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_genotypes.R, R/merge_geno.R
\name{merge_geno}
\alias{merge_geno}
\alias{edit_samples_geno}
\title{Manipulate marker.geno objects}
\usage{
merge_geno(
  geno.list,
  run.ids = NULL,
  sep = ".",
  run.id.append = c("all", "only.duplicated", "none"),
  duplicate.action = c("merge", "keep", "keep.lowest.missing")
)

edit_samples_geno(x, replacement)

merge_geno(
  geno.list,
  run.ids = NULL,
  sep = ".",
  run.id.append = c("all", "only.duplicated", "none"),
  duplicate.action = c("merge", "keep", "keep.lowest.missing")
)
}
\arguments{
\item{geno.list}{A list of data frames of genotype calls, where the first three columns are "marker," "chrom,", and "position."}

\item{run.ids}{A vector of characters defining the run ID for each data frame in \code{geno.list}.
This ID will be added as a suffix in repeated sample ID in case they exist in different files.}

\item{sep}{The string to separate sample names from the suffix in order to make samples unique.}

\item{run.id.append}{When should run.ids be appended? If "all", all samples will have a run.id appeneded; if
"only.duplicated", only duplicated samples will have a run.id appended; if "none", run.ids will not be used to
make sample names unique.}

\item{duplicate.action}{The action to take to handle duplicates between marker genotype data frames. If \code{"merge"},
duplicates will be merged into a single consensus individual; if \code{"keep"}, duplicates will be retained in the final
output data frame; if \code{"keep.lowest.missing"}, the duplicate with the lowest missing data rate will be retained.}

\item{replacement}{A \code{data.frame} with two columns: the sample name as specified in the \code{marker.geno} object; and the corrected # sample name to use.}

\item{...}{Two or more \code{marker.geno} objects.}

\item{lst}{A list of \code{marker.geno} objects.}
}
\value{
A data.frame of merged genotype calls, where the first three columns are "marker," "chrom,", and "position."
}
\description{
Manipulate marker.geno objects

Merge SNP marker genotypes
}
\section{Functions}{
\itemize{
\item \code{edit_samples_geno()}: Edit marker genotype sample names

}}
