% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_vcfs.R
\name{merge_vcfs}
\alias{merge_vcfs}
\title{Merge VCF files together}
\usage{
merge_vcfs(vcf.list)
}
\arguments{
\item{vcf.list}{A list of \code{vcfR} objects.}
}
\value{
A \code{vcfR} object with merged variant data.
}
\description{
Merge VCF files together
}
