% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_summaries.R, R/methods.R
\name{summarize_geno}
\alias{summarize_geno}
\alias{plot,geno.summary-method}
\title{Summarize a \code{geno} object (and derived objects)}
\usage{
summarize_geno(geno)

\S4method{plot}{geno.summary}(x)
}
\arguments{
\item{geno}{A \code{geno} object (or its derivatives).}
}
\description{
Summarize a \code{geno} object (and derived objects)
}
\section{Functions}{
\itemize{
\item \code{plot(geno.summary)}: 

}}
