% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_marker_genotypes.R
\name{vcf2genomat}
\alias{vcf2genomat}
\title{Convert a VCF object to a geno object}
\usage{
vcf2genomat(vcf.in, ploidy = 2, n.core = 1)
}
\arguments{
\item{ploidy}{The ploidy level.}

\item{n.core}{The number of computer cores to use.}

\item{x}{A \code{vcfR} object.}
}
\description{
Convert a VCF object to a geno object
}
